+++ src/api/portfolioApi.js
@@ -1,21 +1,19 @@
 import instance from '@/api/index.js';
-import axios from 'axios';
 
-const portfolioApi = axios.create({
-  baseURL: 'http://localhost:8080/api/', // API 엔드 포인트
-  headers: {
-    'Content-Type': 'application/json',
-  },
-});
+// const portfolioApi = axios.create({
+//     baseURL: 'http://localhost:8080/api/portfolio', // API 엔드 포인트
+//     headers: {
+//         'Content-Type': 'application/json',
+//     },
+// });
 
 export async function fetchPortfolioList() {
   try {
     const response = await instance.get('/portfolio/list');
-    console.log(fetchPortfolioList);
     return response.data;
   } catch (error) {
     console.error('Error fetching portfolio list:', error);
-    throw error;
+    throw error; // 오류 발생 시 오류를 던져줌
   }
 }
 
@@ -28,7 +26,6 @@ export async function getPortfolioDetail(portfolioId) {
     throw error; // 오류 발생 시 오류를 던져줌
   }
 }
-console.log(getPortfolioDetail);
 
 export async function postPortfolio(portfolioReqDto) {
   try {
