+++ src/memo2.vue
@@ -1,0 +1,5 @@
+<template></template>
+
+<script></script>
+
+<style scoped></style>
